Mark internals of `neo4j.spatial` private
